Move components under chen_2009
This paves the way to have the components modules generated from the literate
documents instead.
