--- rasmodel/components/__init__.py
@@ -1,4 +1,0 @@
-"""
-    REM: RAS Executable Model
-    ========================================
-"""
